reorder the flow of tests
